scene: fuzz both deserializers, and six bugs, two memory-unsafe
Task said "scene-deserializer fuzzer" (singular). There are two, and only
fuzzing both is worth anything: `assetlib::loadScene` reads the cooked binary,
`EntitySerde::createEntity` reads the authored JSON, and each had bugs the
other's target could never have found.

── fuzz_scene_loader_test (the cooked binary) ────────────────────────────────
Found the SAME two bugs `fuzz_mesh_loader_test` had already found and fixed in
the mesh loader, carried verbatim:

  - allocation sized straight from a header count. 2.7 GB requested from a
    3 482-byte file; entityCount=0xFFFFFFFF asks for 1 TB. A 20-iteration
    explore run was SIGKILLed by the OOM killer.
  - `return f.good() || f.eof()` accepted truncation as success, handing back
    zero-filled records while the header still claimed the original counts.

That is the finding worth keeping: a hardening that lands in one deserializer
and not its sibling is a coincidence, not a policy. And the scene path is the
more exposed of the two, a shipped dist opens one with no registry and no
source assets to fall back on.

A third was unique to it: `stringTableRead` computed `offset + length` in
32-bit. offset=0xFFFFFFF0 with length=0x20 wraps to 0x10, passes the bound, and
builds a std::string from a pointer 4 GB past the buffer. Reproduced as an
88-byte read from offset 4294967234 in a 32-byte table.

── fuzz_entity_serde_test (the authored JSON) ────────────────────────────────
A `.scene` is a text file people hand-edit, merge, and resolve conflicts in, so
malformed is the normal case, not the adversarial one. Three bugs, each hidden
behind the one before it:

  - every wrong-typed field THREW. nlohmann's `value()` does not fall back to
    the default when a key exists with the wrong type. `scene_serializer.h`
    wraps only `json::parse` in try/catch, so one `"fov": "60"` propagated out
    of scene load and took the editor or the player with it. ~1 case in 20.
  - `j["position"][0..2]` on a short array is UNDEFINED BEHAVIOUR, nlohmann's
    const `operator[](size_type)` is not `at()`, has no bounds check, and
    returns a reference to nothing. Five sites. Segfault, unwindable stack.
    Only reachable once the throw above stopped getting there first.
  - non-finite values accepted. JSON has no NaN literal, but `1e999` parses to
    +inf, and an infinite scale or fov corrupts a frame nowhere near the load.

Component loads are now individually try/caught, and the granularity is the
fix: a malformed COMPONENT is skipped and logged, the rest of the entity still
loads. Wrapping the function would turn a bad field into a missing entity;
wrapping the load, the obvious fix, would turn it into a missing scene.

All nine seeds are pinned in tests/fuzz/corpus/, and each fix was verified by
reverting it alone: 9 failures without the tolerance, a segfault without the
bounds check, 6 failures without the finite guard.

`src/scene` reaches `hardened`, it had `parses-external-input: true` and no
fuzz test, which was exactly what the tier required. `src/runtime` stays
`working` and the roadmap says why: the fuzzers cover the parse and the
string-table reads SceneService makes, but nothing drives SceneService itself,
and scenes are only one of its untrusted inputs.

+++ modules/assetlib/include/assetlib/scene_asset.h
@@ -185,7 +185,12 @@ inline std::string stringTableReadZ(const std::vector<char>& table,
 inline std::string stringTableRead(const std::vector<char>& table,
                                    uint32_t offset, uint32_t length) {
     if (offset == 0xFFFFFFFF || length == 0) return {};
-    if (offset + length > table.size()) return {};
+    // 64-bit, and it has to be. Both fields are uint32_t, so `offset + length`
+    // computes in 32-bit and WRAPS: offset=0xFFFFFFF0 with length=0x20 yields
+    // 0x10, sails past the bound, and hands back a std::string built from a
+    // pointer 4 GB beyond the buffer. A corrupt cooked scene — a bad disk, a
+    // truncated DDC blob from another machine — is enough to reach it.
+    if ((uint64_t)offset + (uint64_t)length > (uint64_t)table.size()) return {};
     return std::string(table.data() + offset, length);
 }
